requirements.txt
discord.py
asyncio
numba
numexpr
numpy
numpydoc